+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - `ArrayBuilder<T>.ToImmutableArray()` for converting the builder contents to an immutable array
 - `VariableNames.AppendGlobalPrefix` for appending `global::` to a span or string
 - `SyntaxExtensions.HasAttribute(string)` for checking if a symbol has a specific attribute
+- `StringBuilderExtensions.Append(ReadOnlySpan<char>)` for appending a read-only character span to a StringBuilder instance
+- `StringBuilderExtensions.AppendLine(ReadOnlySpan<char>)` for appending a read-only character span with a new-line
 
 ## 1.1.0 - 2026-08-01